backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1

# LangChain
langchain==0.2.6
langchain-community==0.2.6
langchain-groq==0.1.6

# Vector store
faiss-cpu==1.8.0

# PDF parsing
pymupdf==1.24.5

# Free local embeddings
sentence-transformers==3.0.1

# Dev
pytest==8.2.2
httpx==0.27.0
